+++ src/QuickMarcEditor/QuickMarcEditorRows/FixedField/MixedMaterialFixedField.js
@@ -19,6 +19,7 @@ const config = {
     {
       name: 'Lang',
       type: SUBFIELD_TYPES.STRING,
+      length: 3,
     },
     {
       name: 'Form',
@@ -31,6 +32,7 @@ const config = {
     {
       name: 'Ctry',
       type: SUBFIELD_TYPES.STRING,
+      length: 3,
     },
     {
       name: 'DtSt',
@@ -39,10 +41,12 @@ const config = {
     {
       name: 'Date1',
       type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
     {
       name: 'Date2',
       type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
   ],
 };