+++ src/QuickMarcEditor/QuickMarcEditorRows/FixedField/ScoreFixedField.js
@@ -23,6 +23,7 @@ const config = {
     {
       type: SUBFIELD_TYPES.STRING,
       name: 'Lang',
+      length: 3,
     },
     {
       type: SUBFIELD_TYPES.BYTE,
@@ -31,6 +32,7 @@ const config = {
     {
       name: 'Comp',
       type: SUBFIELD_TYPES.STRING,
+      length: 2,
     },
     {
       type: SUBFIELD_TYPES.BYTES,
@@ -44,6 +46,7 @@ const config = {
     {
       type: SUBFIELD_TYPES.STRING,
       name: 'Ctry',
+      length: 3,
     },
     {
       type: SUBFIELD_TYPES.BYTE,
@@ -69,10 +72,12 @@ const config = {
     {
       name: 'Date1',
       type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
     {
-      type: SUBFIELD_TYPES.STRING,
       name: 'Date2',
+      type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
   ],
 };